requirements.txt
# Core
boto3>=1.34.0
awsiotsdk>=1.21.0
awscrt>=0.20.0
opencv-python-headless>=4.9.0
pyttsx3>=2.90
PyYAML>=6.0.1
numpy>=1.26.0

# Optional: offline fallback
# ultralytics>=8.0.0   # YOLOv8 (install separately on Pi)

# Observability
prometheus-client>=0.20.0

# Dev / test
pytest>=8.0.0
pytest-cov>=5.0.0
moto[iot,rekognition,s3,dynamodb]>=5.0.0
